get unittest coverage back to 100%
